- added confirmation to removing bbcodes - added optional MX and DNSBL checks - added backtrace (triggering sql error) on error within sql_in_set as well as making sure it is handling an array - let users having f_list access to a forum actually see the forum without a topic list and not displaying an error message - this allows for giving people access to subforums but not the parent forum without the need to add the (sub-)forum to the index. - some additional bugfixes
git-svn-id: file:///svn/phpbb/trunk@6414 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/acp/acp_bbcodes.php
@@ -208,8 +208,20 @@ class acp_bbcodes
 
 				if ($row)
 				{
-					$db->sql_query('DELETE FROM ' . BBCODES_TABLE . " WHERE bbcode_id = $bbcode_id");
-					add_log('admin', 'LOG_BBCODE_DELETE', $row['bbcode_tag']);
+					if (confirm_box(true))
+					{
+						$db->sql_query('DELETE FROM ' . BBCODES_TABLE . " WHERE bbcode_id = $bbcode_id");
+						add_log('admin', 'LOG_BBCODE_DELETE', $row['bbcode_tag']);
+					}
+					else
+					{
+						confirm_box(false, $user->lang['CONFIRM_OPERATION'], build_hidden_fields(array(
+							'bbcode'	=> $bbcode_id,
+							'i'			=> $id,
+							'mode'		=> $mode,
+							'action'	=> $action))
+						);
+					}
 				}
 
 			break;
